requirements.txt
langchain-groq
langchain-mcp-adapters
mcp
python-dotenv
langchain-openai
langchain-core
httpx
streamlit
requests
pandas
urllib3